Fix modifying the global syntax table
This commit fixes modifying syntax entries to the standard global syntax
table.  The previous situation, modifications to the global syntax
table, broke other modes, such as org-mode when tangling with command
org-babel-tangle.  It also changed the meaning of command
isearch-forward-regexp with regular expression '\s-' or '\S-'.  This fix
modifies the syntax entries on the created syntax table.