Add autoloads to minor mode and server commands (#20)
Sometimes when I start editing a new Pollen project, I'll work on the
template first, which would ideally be in web-mode +
pollen-minor-mode. However, as pollen-minor-mode isn't autoloaded, I'd
have to M-: (require 'pollen-mode) manually before being able to use
it. The story is similar with pollen-server-start.

Adding autoload cookies on pollen-minor-mode and pollen-server-start
fixes this problem.