Region (#23)
* make insert react on region

* fix markdown typo (#21)

* add autoload to auto-mode-alist

The autoload is not marked around the file suffix mapping.  Thus
pollen-mode is not activated automatically when the pm file is opened
the first time after Emacs is restarted.  Users have to manually call
pollen-mode to trigger the autoload by file suffix.

Adding autoload to auto-mode-alist fixes the issue.

* update README for autoload

* update commentary section

The old commentary contains comments not useful to the end users.
Pull those comments into the code section.

* make insert react on region

* update commentary section

The old commentary contains comments not useful to the end users.
Pull those comments into the code section.

Co-authored-by: bc² <odanoburu@users.noreply.github.com>